+++ packages/json-rpc/json-rpc-provider-proxy/src/subscription-manager/chainHeadFollow.ts
@@ -58,6 +58,7 @@ export const chainHeadFollow = (
             subscription: id,
             result: {
               event: STOP_EVENT,
+              eventType: "internal",
             },
           },
         }),