+++ packages/client/CHANGELOG.md
@@ -6,12 +6,16 @@
 
 - `polkadot-sdk-compat`: Add `fixChainSpec` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5539) on the PolkadotSDK node.
 - `polkadot-sdk-compat`: Add `fixDescendantValues` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5589) on the PolkadotSDK node.
+- `polkadot-sdk-compat`: Add `fixDescendantValues` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5589) on the PolkadotSDK node.
+- `polkadot-sdk-compat`: Add `fixPrematureBlocks` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5761) on the PolkadotSDK node.
 - `client`: Export `CompatibilityToken` type
 
 ### Fixed
 
 - codegen: reduced startup memory usage due to esbuild issue [#711](https://github.com/polkadot-api/polkadot-api/pull/711)
 - `polkadot-sdk-compat`: Fix small bug with the unpin-hash enhancer
+- `json-rpc-proxy`: Added a property to identify internal "stop" events.
+- `ws-provider`: Some logs when there is a transport close/error
 
 ## 1.2.1 - 2024-09-10